+++ internal/config/config.go
@@ -14,7 +14,7 @@ type Config struct {
 	LogLevel string `yaml:"log_level" env:"BOT_LOG_LEVEL" env-default:"info"`
 	// your bot token as given by botfathers
 	BotToken   string   `yaml:"bot_token" env:"BOT_TOKEN" env-required:"true"`
-	Recepients []string `yaml:"recepients" env:"BOT_RECEPIENTS" env-required:"true"`
+	Recipients []string `yaml:"recipients" env:"BOT_RECIPIENTS" env-required:"true"`
 	Address    string   `yaml:"address" env:"BOT_ADDR" env-default:"localhost:6000"`
 	// API key, passed in 'x-api-key' to authorize requests to the app
 	ApiKey string `yaml:"api_key" env:"BOT_API_KEY" env-required:"true"`
@@ -31,8 +31,8 @@ func MustLoad(configPath string) *Config {
 		log.Fatalf("Error loading configuration %s: %s", configPath, err)
 	}
 
-	if len(cfg.Recepients) == 0 {
-		log.Fatal("No recepients were provided in the config, operation is impossible")
+	if len(cfg.Recipients) == 0 {
+		log.Fatal("No recipients were provided in the config, operation is impossible")
 	}
 
 	if l := len(cfg.ApiKey); l < 60 {