+++ internal/http/handlers/healthcheck.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 
 	"github.com/religiosa1/tgnotifier"
-	"github.com/religiosa1/tgnotifier/internal/http/middleware"
 )
 
 type Healthcheck struct {
@@ -12,8 +11,7 @@ type Healthcheck struct {
 }
 
 func (h Healthcheck) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logger := middleware.GetLogger(r.Context())
-	if _, err := h.Bot.GetMe(logger); err != nil {
+	if _, err := h.Bot.GetMeWithContext(r.Context()); err != nil {
 		w.WriteHeader(http.StatusInternalServerError)
 	}
 	w.WriteHeader(http.StatusOK)